Adição e reorganização testes
