+++ Saudox/tests/Feature/Profissional_AnamneseTest.php
@@ -14,7 +14,7 @@ use App\AnamneseGigantePsicopedaNeuroPsicomotoPt1;
 use App\AnamneseGigantePsicopedaNeuroPsicomotoPt2;
 use App\AnamneseGigantePsicopedaNeuroPsicomotoPt3;
 
-class AnamneseTest extends TestCase {
+class ProfissionalAnamneseTest extends TestCase {
     public $profissional;
     private $endereco;
     private $paciente;
@@ -41,204 +41,6 @@ class AnamneseTest extends TestCase {
             'id_paciente' => $this->paciente->id,
             'id_profissional' => $this->profissional->id,
         ]);
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacientePodeVerPaginaAnamnesePacienteSeEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-
-        $resposta = $this->post(route("login"), [
-
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        //$this->visit(route("paciente.anamnese"));
-        //$this->seePageIs(route("paciente.anamnese"));
-
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacientePodeVerAnamnesePsicoNeuroMotoPacienteSeEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-        factory(AnamneseGigantePsicopedaNeuroPsicomoto::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt1::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt2::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt3::class)->create();
-
-
-        $resposta = $this->post(route("login"), [
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        //$this->visit(route("paciente.anamnese.neuropsicomotora"));
-        //$this->seePageIs(route("paciente.anamnese.neuropsicomotora"));
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacientePodeVerAnamneseFonoaudiologicaPacienteSeEstiverLogado() {
-
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-        factory(AnamneseFonoaudiologia::class)->create([
-            'id_paciente' => $paciente->id,
-            'id_profissional' => $this->profissional->id,
-        ]);
-
-        $resposta = $this->post(route("login"), [
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        //$this->visit(route("paciente.anamnese.fonoaudiologia"));
-        //$this->seePageIs(route("paciente.anamnese.fonoaudiologia"));
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacientePodeVerAnamneseTerapiaOcupacionalPacienteSeEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-        factory(AnamneseTerapiaOcupacional::class)->create([
-            'id_paciente' => $paciente->id,
-            'id_profissional' => $this->profissional->id,
-        ]);
-
-        $resposta = $this->post(route("login"), [
-
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        //$this->visit(route("paciente.anamnese.terapia_ocupacional"));
-        //$this->seePageIs(route("paciente.anamnese.terapia_ocupacional"));
-
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacienteNaoPodeVerAnamnesePsicologiaPacienteSeNaoEstiverLogado() {
-
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-
-        $resposta = $this->post(route("login"), [
-
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-        factory(AnamneseGigantePsicopedaNeuroPsicomoto::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt1::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt2::class)->create();
-        factory(AnamneseGigantePsicopedaNeuroPsicomotoPt3::class)->create();
-
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        $this->post(route("paciente.logout"));
-
-        //$this->visit(route("paciente.anamnese.neuropsicomotora"));
-        //$this->seePageIs(route("login"));
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacienteNaoPodeVerAnamneseTerapiaOcupacionalPacienteSeNaoEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-        $resposta = $this->post(route("login"), [
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-        factory(AnamneseTerapiaOcupacional::class)->create([
-            'id_paciente' => $paciente->id,
-            'id_profissional' => $this->profissional->id,
-        ]);
-
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        $this->post(route("paciente.logout"));
-
-        //$this->visit(route("paciente.anamnese.terapia_ocupacional"));
-        //$this->seePageIs(route("login"));
-
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacienteNaoPodeVerAnamneseFonoaudiologiaPacienteSeNaoEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-
-        $resposta = $this->post(route("login"), [
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-
-        factory(AnamneseFonoaudiologia::class)->create([
-            'id_paciente' => $paciente->id,
-            'id_profissional' => $this->profissional->id,
-        ]);
-
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        $this->post(route("paciente.logout"));
-
-        //$this->visit(route("paciente.anamnese.fonoaudiologia"));
-        //$this->seePageIs(route("login"));
-    }
-
-    /** @ test **/
-    /* url: https://www.pivotaltracker.com/story/show/174638550 */
-    /* TA_01 */
-    public function pacienteNaoPodeVerPaginaAnamnesePacienteSeNaoEstiverLogado() {
-        $paciente = factory(Paciente::class)->create([
-            'password' => bcrypt($password = '123123123'),
-        ]);
-
-        $resposta = $this->post(route("login"), [
-            'login' => $paciente->login,
-            'password' => $password,
-        ]);
-
-        $resposta->assertRedirect(route("paciente.home"));
-        $this->assertAuthenticatedAs($paciente);
-
-        $this->post(route("paciente.logout"));
-
-        //$this->visit(route("paciente.anamnese"));
-        //$this->seePageIs(route("login"));
     }
 
     /** @ test **/
